voice-conversion/requirements.txt
torch
fire
pyyaml
librosa
webrtcvad
numpy
matplotlib
scipy
soundfile
tensorboardX
pysptk
h5py